requirements.txt
PyYAML>=6.0
lunardate>=0.2.2
